Add division slider with limits
Magic numbers have been made as part of the class constructor.  You
might protest, "You should not have to use a constructor!  This is
Traits, where one can use it's MyClass(foo=bar) magic initialization".
The problem with that conventional approach, such as:

    class MicropointAttenuator(HasTraits):
        divisions = 90
        division = Range(1, divisions, value=1)
        ...

    mp = MicropointAttenuator(divisions=30)
    mp.configure_traits()

...is the UI does not initialize correctly.  The range slider displays
a max value of 90, even though the internal Range logic respects the
value of 30.